fix: return a line that was deleted in #49
A line of code the previous commit (#49) was deleted, but in
this way the methods were called from another object (line 934),
although with the same name. By returning the line, you can now run
the demo on Windows.

The original problem was name shadowing. I solved it by renaming
the object.