feat(ci): add the build stage (#69)
* feat(ci): add build stage

The commit adds a project build step to CI in addition to the CMake
configuration step.

It was also decided that it didn't make sense to do debug builds.

* fix(ci): install 'glslc' with installing full SDK

Unfortunately, I couldn't find a quick way to download 'glslc' without
installing the full SDK, so I decided to download the entire VulkanSDK.

I also updated the VulkanSDK version, because the old one can't be
downloaded anymore.